Add files via upload
all of these are random  images some of which are the logos created for this specific website. before replacing the image with another one be careful of the format of the image whether it is 'jpg' or 'png'. As i said this is my first project and i didnt think through about the details of the stuff.